backend/requirements.txt
# Production Dependencies
alembic==1.13.2
asyncpg==0.29.0
bcrypt==4.2.0
dynaconf==3.2.6
fastapi==0.113.0
httpx==0.27.2
jinja2==3.1.4
markdown==3.7
numpy==1.26.4
pandas==2.2.2
pgvector==0.3.2
pillow==10.4.0
psutil==6.0.0
psycopg2-binary==2.9.9
pydantic==2.9.0
pydantic-settings==2.4.0
pygments==2.18.0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
pyyaml==6.0.2
requests==2.31.0
schedule==1.2.2
scikit-learn==1.5.1
sentence-transformers==3.0.1
sqlalchemy==2.0.34
tiktoken==0.7.0
toml==0.10.2
torch==2.2.1
transformers==4.38.2
tzlocal==5.2
uuid7==0.1.0
uvicorn==0.30.0

# Development Dependencies
bandit==1.7.9
black==24.2.0
coverage==7.4.3
flake8==7.0.0
ipython==8.22.2
mypy==1.8.0
pip-tools==7.4.1
pre-commit==3.6.2
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
time-machine==2.13.0
watchfiles==0.21.0
